chore(trie): move hash builder to primitives
