+++ crates/primitives/src/trie/mod.rs
@@ -1,17 +1,21 @@
 //! Collection of trie related types.
 
-mod branch_node;
-mod hash_builder;
+/// Various branch nodes produced by the hash builder.
+pub mod nodes;
+pub use nodes::BranchNodeCompact;
+
+/// The implementation of hash builder.
+pub mod hash_builder;
+pub use hash_builder::HashBuilder;
+
 mod mask;
 mod nibbles;
 mod storage;
 mod subnode;
 
 pub use self::{
-    branch_node::BranchNodeCompact,
-    hash_builder::{HashBuilderState, HashBuilderValue},
     mask::TrieMask,
-    nibbles::{StoredNibbles, StoredNibblesSubKey},
+    nibbles::{Nibbles, StoredNibbles, StoredNibblesSubKey},
     storage::StorageTrieEntry,
     subnode::StoredSubNode,
 };
