+++ crates/trie/src/lib.rs
@@ -9,17 +9,8 @@
 //! authenticated radix trie that is used to store key-value bindings.
 //! <https://ethereum.org/en/developers/docs/data-structures-and-encoding/patricia-merkle-trie/>
 
-mod nibbles;
-pub use nibbles::Nibbles;
-
 /// The Ethereum account as represented in the trie.
 pub mod account;
-
-/// Various branch nodes produced by the hash builder.
-pub mod nodes;
-
-/// The implementation of hash builder.
-pub mod hash_builder;
 
 /// The implementation of a container for storing intermediate changes to a trie.
 /// The container indicates when the trie has been modified.
