+++ crates/trie/src/updates.rs
@@ -1,4 +1,3 @@
-use crate::Nibbles;
 use derive_more::Deref;
 use reth_db::{
     cursor::{DbCursorRO, DbCursorRW, DbDupCursorRO, DbDupCursorRW},
@@ -6,7 +5,7 @@ use reth_db::{
     transaction::{DbTx, DbTxMut},
 };
 use reth_primitives::{
-    trie::{BranchNodeCompact, StorageTrieEntry, StoredNibbles, StoredNibblesSubKey},
+    trie::{BranchNodeCompact, Nibbles, StorageTrieEntry, StoredNibbles, StoredNibblesSubKey},
     H256,
 };
 use std::collections::{hash_map::IntoIter, HashMap};
